requirements.txt
aws-cdk-lib==2.145.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-amplify-alpha==2.149.0a0